backend/pyproject.toml
[tool.poetry]
name = "ddialog-backend"
version = "0.1.0"
description = "Backend for DDialog application"
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"